moved anything used on all routes to router
