+++ routes/users.php
@@ -1,13 +1,5 @@
 <?php
-require_once('utils/preflight-check.php');
-
-header('Access-Control-Allow-Origin: ' . $_SERVER['HTTP_ORIGIN']);
-header('Access-Control-Allow-Headers: Content-Type, Authorization');
-header('Access-Control-Allow-Methods: GET, POST, PUT, DELETE, OPTIONS');
-header('Access-Control-Allow-Credentials: true');
-
 require_once('utils/authorize.php');
-require_once('utils/database-connection.php');
 
 $input = json_decode(file_get_contents('php://input'), true);
